+++ src/chroma/despawner.ts
