+++ src/chroma/laserScene.ts
