+++ src/chroma/shapeGenerator.ts
