+++ src/functions/logMapData.ts
