+++ src/functions/random.ts
