+++ src/noodle/animatePlayer.ts
