+++ src/noodle/notePath.ts
